[f] update style for No Internet Connection Banner
