+++ BitmarkRegistry/Services/Global.swift
@@ -22,11 +22,13 @@ class Global {
   static var currentJwt: String?
   static var apnsToken: String? // Push notification token
   static var verificationLink: String? // Chibitronics
-  static var noInternetBanner: NotificationBanner = {
-    return NotificationBanner(
-      title: "", subtitle: "NoInternetConnection".localized().localizedUppercase,
+  static var noInternetBanner: StatusBarNotificationBanner = {
+    let statusBarNotificationBanner = StatusBarNotificationBanner(
+      title: "NoInternetConnection".localized(),
       style: .danger, colors: CustomBannerColors()
     )
+    statusBarNotificationBanner.applyStyling(titleFont: UIFont(name: "Avenir-Black", size: 12)!)
+    return statusBarNotificationBanner
   }()
 
   public static func clearData() {
@@ -111,23 +113,10 @@ class Global {
 extension Global {
   static func showNoInternetBanner() {
     noInternetBanner.show()
-    noInternetBanner.applyStyling(subtitleFont: UIFont(name: "Avenir-Black", size: 16), subtitleTextAlign: .center)
   }
 
   static func hideNoInternetBanner() {
     noInternetBanner.dismiss()
-  }
-
-  public class CustomBannerColors: BannerColorsProtocol {
-    public func color(for style: BannerStyle) -> UIColor {
-      switch style {
-      case .danger:   return UIColor.mainRedColor
-      case .info:     return UIColor(red:0.23, green:0.60, blue:0.85, alpha:1.00)
-      case .customView:     return UIColor.clear
-      case .success:  return UIColor(red:0.22, green:0.80, blue:0.46, alpha:1.00)
-      case .warning:  return UIColor(red:1.00, green:0.66, blue:0.16, alpha:1.00)
-      }
-    }
   }
 }
