+++ BitmarkRegistry/Supporting Files/en.lproj/Localizable.strings
@@ -35,7 +35,7 @@
 "Logout »" = "Log Out »";
 "NeedHelp?" = "Need Help?";
 "No" = "No";
-"NoInternetConnection" = "NO INTERNET CONNECTION!";
+"NoInternetConnection" = "No Internet Connection";
 "OK" = "OK";
 "Owner" = "Owner";
 "P2P_TRANSFER" = "P2P Transfer";
