+++ BitmarkRegistry/Supporting Files/zh-Hant.lproj/Localizable.strings
@@ -35,7 +35,7 @@
 "Logout »" = "登出 »";
 "NeedHelp?" = "需要幫忙？";
 "No" = "取消";
-"NoInternetConnection" = "無網路連線！";
+"NoInternetConnection" = "無網路連線";
 "OK" = "確認";
 "Owner" = "所有人";
 "P2P_TRANSFER" = "轉移";
